+++ src/generic-table/interfaces/gt-options.ts
@@ -1,4 +1,5 @@
 export interface GtOptions {
   cache: boolean;
   debounceTime: number;
+  csvDelimiter: string;
 }